==> ImmuneCellAnalysis/tenX/Immune/01a_Preprocessing.Rmd <==
---
title: "R Notebook"
output: html_document
---

## Preprocess matrices to seurat object, then do broad annotation (myeloid, tcell) + detailed (cd4, cd8) and save these seurat objects

```{r}
knitr::opts_chunk$set(echo = TRUE)

base_dir = "/Users/jlabelle/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/ImmuneCellAnalysis/tenX/Immune/")
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Shared/Scripts/"


preprocessing = paste0(base_dir, script_dir, "single_cell_preprocessing_helper_functions.R")
source(preprocessing)
library(Matrix)

analysis_dir<- "analysis/preprocessing/"
figure_dir<- "figures/preprocessing/"
data_dir<- "data/"

if(!dir.exists(analysis_dir)){dir.create(analysis_dir, recursive = TRUE)}
if(!dir.exists(figure_dir)){dir.create(figure_dir, recursive = TRUE)}

marker_dir<- paste0(base_dir, "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Shared/Marker_genes/")
hkgenes <- read.table(paste0(marker_dir, "tirosh_house_keeping.txt"), skip = 2)
hkgenes <- as.vector(hkgenes$V1)
```

## Read in immune matrixes- 1 per sample
```{r}
samples<- list.files(data_dir)

all_matrix<- lapply(samples, function(x){
  print(x)
  matrix_dir<- list.files(paste0(data_dir, x))
  matrix_dir<- matrix_dir[!grepl("VDJ", matrix_dir) & !grepl("TCR", matrix_dir)]
  
  ## Load matrices for each dir- BT1857 has 2 tenx samples demultiplexed separately. BT2082/2083 have 1 each
  ## Load in 2 BT1857 matrices, merge into single df
  if(length(matrix_dir) > 1){
    sample_matrix_list<-lapply(matrix_dir, function(y){
      print(y)
      sample_matrix_dir<- paste0(data_dir, x, "/", y, "/filtered_feature_bc_matrix/")
      
      barcode.path <- paste0(sample_matrix_dir, "barcodes.tsv.gz")
      features.path <- paste0(sample_matrix_dir, "features.tsv.gz")
      matrix.path <- paste0(sample_matrix_dir, "matrix.mtx.gz")
      mat <- readMM(file = matrix.path)
      feature.names = read.delim(features.path,
                                 header = FALSE,
                                 stringsAsFactors = FALSE)
      barcode.names = read.delim(barcode.path,
                                 header = FALSE,
                                 stringsAsFactors = FALSE)
      colnames(mat) = barcode.names$V1
      rownames(mat) = feature.names$V2
      
      ## Add sample ID to matrix
      colnames(mat)<- paste0(y, "_", colnames(mat))
      return(mat)
    })
    names(sample_matrix_list)<- matrix_dirs
    
    ## merge into single df
    sum(rownames(sample_matrix_list[[1]]) != rownames(sample_matrix_list[[2]]) )
    sample_matrix_df<- do.call("cbind", sample_matrix_list)
    
    ## read in BT2082/BT2083 separately
  } else{
    print(x)
    sample_matrix_dir<- paste0(data_dir, x, "/", matrix_dir, "/filtered_feature_bc_matrix/")
    barcode.path <- paste0(sample_matrix_dir, "barcodes.tsv.gz")
    features.path <- paste0(sample_matrix_dir, "features.tsv.gz")
    matrix.path <- paste0(sample_matrix_dir, "matrix.mtx.gz")
    mat <- readMM(file = matrix.path)
    feature.names = read.delim(features.path,
                               header = FALSE,
                               stringsAsFactors = FALSE)
    barcode.names = read.delim(barcode.path,
                               header = FALSE,
                               stringsAsFactors = FALSE)
    colnames(mat) = barcode.names$V1
    rownames(mat) = feature.names$V2
    
    ## Add sample ID to matrix
    colnames(mat)<- paste0(x, "_", colnames(mat))
    sample_matrix_df<- mat
    
    ## replace all - with _ in column names
    colnames(mat)<- gsub("-", "_", colnames(mat))
  }
  return(sample_matrix_df)
})
names(all_matrix)<- samples
```

## Make simple metadata for each matrix- cells + sample
```{r}
all_meta<- lapply(names(all_matrix), function(x){
  matrix<- all_matrix[[x]]
  meta<- data.frame(cell=colnames(matrix), sample=rep(x, ncol(matrix)))
  return(meta)
})
```

## Save cm and meta
```{r}
meta<- do.call("rbind", all_meta)

sum(rownames(all_matrix[[1]]) != rownames(all_matrix[[2]]))
cm<- do.call("cbind", all_matrix)

## There are 95 genes that are duplicated (same for both samples)-sum counts
  ## Get cm of genes with dup genes
  dup_genes<- rownames(cm)[duplicated(rownames(cm))]
  dup_cm<-cm[rownames(cm) %in% dup_genes,]
  dup_genes<- rownames(dup_cm)
  dup_cm<- as.data.frame(dup_cm)
  dup_cm$gene<- dup_genes

  ## Group by gene and get total counts across cell
  dup_cm<- dup_cm %>% group_by(gene) %>% summarise(across(everything(), list(sum))) %>% as.data.frame()
  rownames(dup_cm)<-dup_cm$gene
  dup_cm$gene<- NULL
  colnames(dup_cm)<- gsub("_1", "", colnames(dup_cm))
  
  ## Add back to full cm
  cm_unDup<- as.data.frame(cm[!rownames(cm) %in% rownames(dup_cm),])
  cm<- rbind(cm_unDup, dup_cm)

saveRDS(cm, file=paste0(analysis_dir, "raw_cm.Rds"))
saveRDS(meta, file=paste0(analysis_dir, "raw_meta.Rds"))
```

## Create seurat object, do initial QC
```{r}
cm<- readRDS(paste0(analysis_dir, "raw_cm.Rds"))
meta<- readRDS(paste0(analysis_dir, "raw_meta.Rds"))


seurat_raw<- RunFullSeurat(cm=cm, RunHarmony = FALSE, samples=meta$sample)


 ## Identify # of HK genes
hkgenes.notfound.indices <- which(!hkgenes %in% rownames(seurat_raw))  
if (length(hkgenes.notfound.indices)) {
  message = paste0(paste0(hkgenes[hkgenes.notfound.indices], collapse = " "), " housekeeping genes were not found")
  print(message)
}


# Get number of expressed housekeeping genes.
hkgenes.found <- hkgenes[-hkgenes.notfound.indices]  # remove hkgenes that were not found
seurat_raw[["nHK_RNA"]] <- Matrix::colSums(seurat_raw[['RNA']]@data[hkgenes.found, ] > 0)


## Get percentage of mito genes
seurat_raw[["percent_mito"]] <- PercentageFeatureSet(seurat_raw, pattern = "^MT-")
  


## QC plots
p1<-DimPlot(seurat_raw, group.by="sample", label=TRUE, pt.size = 0)+NoLegend()
p2<- VlnPlot(seurat_raw, features = "nFeature_RNA", group.by="sample", pt.size = 0)+NoLegend()
p3<- VlnPlot(seurat_raw, features="percent_mito", group.by="sample", pt.size = 0)+NoLegend()
p4<- VlnPlot(seurat_raw, features="nHK_RNA", group.by="sample", pt.size = 0)+NoLegend()


p1+p2+p3+p4
ggsave(paste0(figure_dir, "nFeature.Count_BySample.png"), width=8, height=10)
saveRDS(seurat_raw, file=paste0(analysis_dir, "raw_seurat.Rds"))
```

## Plot histogram of nFeature/nHK/nMitochondrial to determine good cutoff
```{r}
df<- seurat_raw@meta.data
p1<- ggplot(df, aes(x=nFeature_RNA, fill=sample))+geom_histogram()+ggtitle("nFeatures")
p2<- ggplot(df, aes(x=nHK_RNA, fill=sample))+ geom_histogram()+ggtitle("n HK genes")+theme(legend.position = "none")
p3<- ggplot(df, aes(x=nCount_RNA, fill=sample))+ geom_histogram() + ggtitle("nCounts")
p4<- ggplot(df, aes(x=percent_mito, fill=sample))+ geom_histogram() + ggtitle("percent mito")+theme(legend.position = "none")
p1+p2+p3+p4
ggsave(paste0(figure_dir, "Histogram_Features.Counts.HK.Mito.png"), width=10, height=8)
```




## Filter cells by nFeature + nHK
```{r}
seurat_raw<- readRDS(paste0(analysis_dir, "raw_seurat.Rds"))

## Filter 
seurat_obj_filt <- subset(seurat_raw, subset = nFeature_RNA > 500 & nFeature_RNA< 5000 & nHK_RNA > 20)
seurat_raw$PF<- ifelse(colnames(seurat_raw) %in% colnames(seurat_obj_filt), "Pass", "Fail")

df<- seurat_raw@meta.data
df<- as.data.frame(table(df$sample, df$PF))
colnames(df)<- c("Sample", "PF", "nCells")
ggplot(df, aes(x=Sample, y=nCells, fill=PF))+
  geom_bar(stat="identity", position="stack")+
  theme(axis.text.x = element_text(angle=45, hjust=1))+
  ggtitle("nFeature_RNA > 500\nnHK_RNA > 20")+
  xlab("")
ggsave(paste0(figure_dir, "PF_BySample_nFeature.nHK.png"), width=3, height=3)

## recluster without low quality cells
seurat_obj_filt_meta<- seurat_obj_filt@meta.data
seurat_obj_filt<- RunFullSeurat(seurat_obj_filt@assays$RNA@counts, RunHarmony = TRUE, samples=seurat_obj_filt_meta$sample)
seurat_obj_filt<-AddMetaData(seurat_obj_filt, seurat_obj_filt_meta[,!colnames(seurat_obj_filt_meta) %in% colnames(seurat_obj_filt@meta.data)])

## Plot by qc features to confirm all low qual removed
FeaturePlot(seurat_obj_filt, features = c("nFeature_RNA", "nHK_RNA", "percent_mito"), cols=c("grey", "red"))
ggsave(paste0(figure_dir, "PostFiltration_Feature.HK.Mito.png"), width=6, height=5)

saveRDS(seurat_obj_filt, file=paste0(analysis_dir, "seurat_obj.Rds"))

```

## Identify and remove any non-immune normal cells (OC, AC)
```{r}
seurat_obj<- readRDS(paste0(analysis_dir, "seurat_obj.Rds"))

## Reduce resolution to identify markers for major clusters
seurat_obj<-FindClusters(seurat_obj, resolution = 0.1)
all_markers<- FindAllMarkers(seurat_obj, min.pct = 0.4, logfc.threshold = 0.5)
all_markers_filtered<- all_markers[all_markers$p_val_adj<0.05,]
top_markers<- all_markers_filtered %>% group_by(cluster) %>% top_n(n=10, wt=avg_log2FC) %>% as.data.frame()

## Assign cluster 6 to AC, cluster r to Unknown (not T cells or myeloid)
seurat_obj$broadest_annot<- ifelse(seurat_obj$seurat_clusters==5, "DC",
                                   ifelse(seurat_obj$seurat_clusters %in% c(4,3), "Tumor", "Immune"))
FeaturePlot(seurat_obj, features=c("CD3G", "CSF1R", "GFAP", "SOX3", "IL3RA")) + DimPlot(seurat_obj, group.by = "broadest_annot")
ggsave(paste0(figure_dir, "UMAP_removeTumor.DC.png"), width=7, height=6)


## Remove tumor/dc
seurat_obj<- subset(seurat_obj, broadest_annot =="Immune")

## Convert scale to TPM- to match with SS2 
cm<- seurat_obj@assays$RNA@counts
cm<- apply(cm, 2, function(x){x/sum(x) * 1e06})

## Rerun clustering (just so processing is completely identical to ss2)
seurat_obj<- RunFullSeurat(cm, RunHarmony = TRUE, samples=seurat_obj$sample)
seurat_obj<- AddMetaData(seurat_obj, seurat_meta[,!colnames(seurat_meta) %in% colnames(seurat_obj@meta.data)])


saveRDS(seurat_obj, file=paste0(analysis_dir, "seurat_obj_noTumor.Rds"))
```
