==> ImmuneCellAnalysis/tenX/Immune/03_NMF.Rmd <==
---
title: "NMF"
author: "Jenna LaBelle"
output: html_document
---

De novo NMF run on tenX untreated samples (n=3) for T cells and myeloid cells. Record of de novo NMF (run on O2) and then finish processing, identification of marker genes, for all ranks

```{r}
library(Seurat) 
library(clusterProfiler)
library(biomaRt)
library(stringr)

base_dir = "/Users/jlabelle/"
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Shared/Scripts/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/ImmuneCellAnalysis/tenX/Immune/")

analysis_dir<- paste0(working_dir, "/analysis/nmf/")
figure_dir<- paste0(working_dir, "/figures/nmf/")
data_dir <- paste0(working_dir, "analysis/nmf/")

if(!dir.exists(figure_dir)){dir.create(figure_dir,recursive = TRUE)}
if(!dir.exists(analysis_dir)){dir.create(analysis_dir, recursive = TRUE)}

preprocessing = paste0(base_dir, script_dir, "single_cell_preprocessing_helper_functions.R")
nmf_helper = paste0(base_dir, script_dir, "NMF_helper_function.R")
source(preprocessing)
source(nmf_helper)
source(paste0(base_dir, script_dir, "plotutils.R"))
source(paste0(base_dir, script_dir,  "seurat3utils.R"))
source(paste0(base_dir, script_dir, "color.R"))


```

## Load in tenx seurat object, de novo nmf for 10x, full cohort programs
```{r}
## Seurat
myeloid_seurat<- readRDS("analysis/manual_broad_annot/myeloid_seurat.Rds")
tcell_seurat<- readRDS("analysis/manual_broad_annot/tcell_seurat.Rds")


## NMF
tcell_estim.r<- readRDS(paste0(data_dir, "tcell_estimR_rank6-10.Rds"))
myeloid_estim.r<- readRDS(paste0(data_dir, "myeloid_estimR_rank10-13.Rds"))
```

## Clustering using non-negative matrix factorization (NMF)
Cluster data using NMF with a range of ranks. The rank is the number of metagenes and the number of clusters. The choice of the rank is heuristic. For each clustering, we also calculate DE genes.

## This section run on O2 for both tcells/myeloid (separately)
```{r}
# Vector of ranks for which to calculate NMF.
rank <- 6:10

# Identify variable genes within this cell subset.
var.genes <- SelectIntegrationFeatures(SplitObject(tcell_seurat, split.by = "sample"), nfeatures = 2000, verbose = TRUE,
                                       fvf.nfeatures = 2000, selection.method = "vst")

VariableFeatures(tcell_seurat) <- var.genes

# Number of variable genes
print(length(VariableFeatures(tcell_seurat)))

# Perform 10 NMF runs for each value of rank r.
x <- as.matrix(tcell_seurat[['RNA']][VariableFeatures(tcell_seurat), ])
x <- x[which(rowSums(x) > 0), ]  # remove any rows filled with 0 (just in case these exist)

#estim.r <- nmf(x, rank = rank, method = "brunet", nrun = 10, seed = 123456)
tcell_estim.r <- nmf(x, rank = rank, method = "brunet", nrun = 10, seed = 123456, .options = "tvP4")

saveRDS(tcell_estim.r, file=paste0(analysis_dir, "estimR_rank6-10.Rds"))
```

## Finish processing for de novo NMF: Tcells
```{r NMF, eval = F}
analysis_dir_use<- paste0(analysis_dir, "Tcells/")
figure_dir_use<- paste0(figure_dir, "Tcells/")

png(paste0(figure_dir_use, "NMF_plot_rank6-10.png"), width = 2000, height = 1000, res = 100)
plot(tcell_estim.r)
dev.off()

# Iterate over different ranks and look at clustering.
rank <- 6:10
for (i in seq(rank)) {
  r <- rank[i]
  print(r)
  
  # Plot coefficient values of metagenes for each cell.
  png(paste0(figure_dir_use, "NMF_coefmap_rank", r, ".png"), width = 2000, height = 1000, res = 100)
  coefmap(tcell_estim.r$fit[[i]], Colv=T)
  dev.off()

  # Show genes contributing the most to each metagene (factor). 
  # Remember, this is limited to the variable genes if that was the input to NMF.
  for (j in seq(r)) {
    top.features <- names(sort(tcell_estim.r$fit[[i]]@fit@W[, j], decreasing = T)[1:10])
    print(paste(top.features, collapse = " "))
  }
  s <- extractFeatures(tcell_estim.r$fit[[i]], 10L)
  for (j in seq(r)) {
    top.features <- rownames(tcell_estim.r$fit[[i]]@fit@W)[s[[j]]]
    print(paste(top.features, collapse = " "))
  }

  # Plot NMF residuals.
  png(paste0(figure_dir_use, "NMF_residuals_rank", r, ".png"), width = 2000, height = 1000, res = 100)
  plot(tcell_estim.r$fit[[i]])
  dev.off()

  # Check coefficient matrix has cell names in same order as Seurat object tcell_seurat.
  H <- tcell_estim.r$fit[[i]]@fit@H
  colnames(H)<- gsub("-", "_", colnames(H))
  all.equal(as.integer(as.character(predict(tcell_estim.r$fit[[i]], what = "samples"))), as.integer(apply(H, 2, which.max)))
  all.equal(colnames(H), colnames(tcell_seurat))
  all.equal(rownames(tcell_estim.r$fit[[i]]@fit@W), VariableFeatures(tcell_seurat))
  
  # Cluster cells based on maximum coefficient for metagene.
  tcell_seurat[[paste0("NMF_rank", r)]] <- predict(tcell_estim.r$fit[[i]], what = "samples")
  Idents(tcell_seurat) <- paste0("NMF_rank", r)

  # UMAP visualization and tcell and tcell expression across clusters and find DE genes.
  DimPlot(tcell_seurat, reduction = "umap", cols = material.heat(n_distinct(Idents(tcell_seurat))), label = T)
  ggsave(paste0(figure_dir_use,  "UMAP_NMF_rank", r, ".png"), width = 6, height = 6)
  VlnPlot(tcell_seurat, features = c("CSF1R", "CD3G"), cols = material.heat(n_distinct(Idents(tcell_seurat))), 
          ncol = 1, pt.size = 0.5) 
  ggsave(paste0(figure_dir_use, "vln_NMF_rank", r, ".png"), width = 8, height = 8)
  
  # Find differentially expressed genes for each cluster.
  tcell_seurat.markers <- FindAllMarkers(tcell_seurat, min.pct = 0.1, logfc.threshold = 0.5, 
                                           max.cells.per.ident = 1500, test.use ="t", 
                                           return.thresh = 0.1 / nrow(tcell_seurat), verbose = T)
  top10genes <- tcell_seurat.markers %>% filter(avg_log2FC > 0) %>% group_by(cluster) %>% top_n(10, avg_log2FC)
  top100genes <- tcell_seurat.markers %>% filter(avg_log2FC > 0) %>% group_by(cluster) %>% top_n(100, avg_log2FC)
  
  # Write cluster markers to csv file.
  write.csv(top100genes, file = paste0(analysis_dir_use,  "top_DE_genes_NMF_rank", r, ".csv"))
  tcell_seurat <- ScaleData(tcell_seurat, features = c(top10genes$gene), do.center = T, do.scale = F)
  DoHeatmap(tcell_seurat, features = top10genes$gene) + scale_fill_distiller(palette = "RdYlBu")
  ggsave(paste0(figure_dir_use,  "DE_genes_expression_rank", r, ".png"), width = 18, height = 24)
}

saveRDS(tcell_seurat, file="analysis/manual_broad_annot/tcell_seurat.Rds")

```

## Finish processing for de novo NMF: Myeloid
```{r NMF, eval = F}
analysis_dir_use<- paste0(analysis_dir, "Myeloid/")
figure_dir_use<- paste0(figure_dir, "Myeloid/")
for (f in c(analysis_dir_use, figure_dir_use)){if(!dir.exists(f)){dir.create(f)}}

png(paste0(figure_dir_use, "NMF_plot_rank10-13.png"), width = 2000, height = 1000, res = 100)
plot(myeloid_estim.r)
dev.off()

# Iterate over different ranks and look at clustering.
## For some reason- rank 8, 10 (in rank6-10) and rank 11 (in rank 10-13) failed??
rank <- 10:13
for (i in seq(rank)) {
  r <- rank[i]
  print(r)
  
  # Plot coefficient values of metagenes for each cell.
  png(paste0(figure_dir_use, "NMF_coefmap_rank", r, ".png"), width = 2000, height = 1000, res = 100)
  coefmap(myeloid_estim.r$fit[[i]], Colv=T)
  dev.off()

  # Show genes contributing the most to each metagene (factor). 
  # Remember, this is limited to the variable genes if that was the input to NMF.
  for (j in seq(r)) {
    top.features <- names(sort(myeloid_estim.r$fit[[i]]@fit@W[, j], decreasing = T)[1:10])
    print(paste(top.features, collapse = " "))
  }
  s <- extractFeatures(myeloid_estim.r$fit[[i]], 10L)
  for (j in seq(r)) {
    top.features <- rownames(myeloid_estim.r$fit[[i]]@fit@W)[s[[j]]]
    print(paste(top.features, collapse = " "))
  }

  # Plot NMF residuals.
  png(paste0(figure_dir_use, "NMF_residuals_rank", r, ".png"), width = 2000, height = 1000, res = 100)
  plot(myeloid_estim.r$fit[[i]])
  dev.off()

  # Check coefficient matrix has cell names in same order as Seurat object myeloid_seurat.
  H <- myeloid_estim.r$fit[[i]]@fit@H
  colnames(H)<- gsub("-", "_", colnames(H))
  all.equal(as.integer(as.character(predict(myeloid_estim.r$fit[[i]], what = "samples"))), as.integer(apply(H, 2, which.max)))
  all.equal(colnames(H), colnames(myeloid_seurat))
  all.equal(rownames(myeloid_estim.r$fit[[i]]@fit@W), VariableFeatures(myeloid_seurat))
  
  # Cluster cells based on maximum coefficient for metagene.
  myeloid_seurat[[paste0("NMF_rank", r)]] <- predict(myeloid_estim.r$fit[[i]], what = "samples")
  Idents(myeloid_seurat) <- paste0("NMF_rank", r)

  # UMAP visualization and myeloid xpression across clusters and find DE genes.
  DimPlot(myeloid_seurat, reduction = "umap", cols = material.heat(n_distinct(Idents(myeloid_seurat))), label = T)
  ggsave(paste0(figure_dir_use,  "UMAP_NMF_rank", r, ".png"), width = 6, height = 6)
  VlnPlot(myeloid_seurat, features = c("CSF1R", "CD3G"), cols = material.heat(n_distinct(Idents(myeloid_seurat))), 
          ncol = 1, pt.size = 0.5) 
  ggsave(paste0(figure_dir_use, "vln_NMF_rank", r, ".png"), width = 8, height = 8)
  
  # Find differentially expressed genes for each cluster.
  myeloid_seurat.markers <- FindAllMarkers(myeloid_seurat, min.pct = 0.1, logfc.threshold = 0.5, 
                                           max.cells.per.ident = 1500, test.use ="t", 
                                           return.thresh = 0.1 / nrow(myeloid_seurat), verbose = T)
  
  top10genes <- myeloid_seurat.markers %>% filter(avg_log2FC > 0) %>% group_by(cluster) %>% top_n(10, avg_log2FC)
  top100genes <- myeloid_seurat.markers %>% filter(avg_log2FC > 0) %>% group_by(cluster) %>% top_n(100, avg_log2FC)
  
  # Write cluster markers to csv file.
  write.csv(top100genes, file = paste0(analysis_dir_use,  "top_DE_genes_NMF_rank", r, ".csv"))
  myeloid_seurat <- ScaleData(myeloid_seurat, features = c(top10genes$gene), do.center = T, do.scale = F)
  DoHeatmap(myeloid_seurat, features = top10genes$gene) + scale_fill_distiller(palette = "RdYlBu")
  ggsave(paste0(figure_dir_use,  "DE_genes_expression_rank", r, ".png"), width = 18, height = 24)
}

saveRDS(myeloid_seurat, file="analysis/manual_broad_annot/myeloid_seurat.Rds")

```
